refactor VideoEntry class
