+++ setup.py
@@ -19,7 +19,8 @@ def create_app() -> Flask:
     app.register_blueprint(views, url_prefix='/')
     app.register_blueprint(auth, url_prefix='/auth')
     app.register_blueprint(videos, url_prefix='/video')
-    from models import User, VideoEntry
+    from models import User
+    from video_entry.video_entry import VideoEntry
 
     create_database(app)
